fix(web): correct type imports for change requests
